// src-old/proto/js/Agent_pb.d.ts
export class EmptyMessage extends jspb.Message { 

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): EmptyMessage.AsObject;
    static toObject(includeInstance: boolean, msg: EmptyMessage): EmptyMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: EmptyMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): EmptyMessage;
    static deserializeBinaryFromReader(message: EmptyMessage, reader: jspb.BinaryReader): EmptyMessage;
}

export namespace EmptyMessage {
    export type AsObject = {
    }
}

export class StringMessage extends jspb.Message { 
    getS(): string;
    setS(value: string): StringMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): StringMessage.AsObject;
    static toObject(includeInstance: boolean, msg: StringMessage): StringMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: StringMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): StringMessage;
    static deserializeBinaryFromReader(message: StringMessage, reader: jspb.BinaryReader): StringMessage;
}

export namespace StringMessage {
    export type AsObject = {
        s: string,
    }
}

export class BooleanMessage extends jspb.Message { 
    getB(): boolean;
    setB(value: boolean): BooleanMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): BooleanMessage.AsObject;
    static toObject(includeInstance: boolean, msg: BooleanMessage): BooleanMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: BooleanMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): BooleanMessage;
    static deserializeBinaryFromReader(message: BooleanMessage, reader: jspb.BinaryReader): BooleanMessage;
}

export namespace BooleanMessage {
    export type AsObject = {
        b: boolean,
    }
}

export class StringListMessage extends jspb.Message { 
    clearSList(): void;
    getSList(): Array<string>;
    setSList(value: Array<string>): StringListMessage;
    addS(value: string, index?: number): string;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): StringListMessage.AsObject;
    static toObject(includeInstance: boolean, msg: StringListMessage): StringListMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: StringListMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): StringListMessage;
    static deserializeBinaryFromReader(message: StringListMessage, reader: jspb.BinaryReader): StringListMessage;
}

export namespace StringListMessage {
    export type AsObject = {
        sList: Array<string>,
    }
}
// ...
export class NewKeyPairMessage extends jspb.Message { 
    getPassphrase(): string;
    setPassphrase(value: string): NewKeyPairMessage;

    getNbits(): number;
    setNbits(value: number): NewKeyPairMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): NewKeyPairMessage.AsObject;
    static toObject(includeInstance: boolean, msg: NewKeyPairMessage): NewKeyPairMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: NewKeyPairMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): NewKeyPairMessage;
    static deserializeBinaryFromReader(message: NewKeyPairMessage, reader: jspb.BinaryReader): NewKeyPairMessage;
}

export namespace NewKeyPairMessage {
    export type AsObject = {
        passphrase: string,
        nbits: number,
    }
}

export class DeriveKeyMessage extends jspb.Message { 
    getKeyName(): string;
    setKeyName(value: string): DeriveKeyMessage;

    getPassphrase(): string;
    setPassphrase(value: string): DeriveKeyMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): DeriveKeyMessage.AsObject;
    static toObject(includeInstance: boolean, msg: DeriveKeyMessage): DeriveKeyMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: DeriveKeyMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): DeriveKeyMessage;
    static deserializeBinaryFromReader(message: DeriveKeyMessage, reader: jspb.BinaryReader): DeriveKeyMessage;
}

export namespace DeriveKeyMessage {
    export type AsObject = {
        keyName: string,
        passphrase: string,
    }
}

export class DeriveKeyPairMessage extends jspb.Message { 

    hasKeypairDetails(): boolean;
    clearKeypairDetails(): void;
    getKeypairDetails(): NewKeyPairMessage | undefined;
    setKeypairDetails(value?: NewKeyPairMessage): DeriveKeyPairMessage;

    getPublicKeyPath(): string;
    setPublicKeyPath(value: string): DeriveKeyPairMessage;

    getPrivateKeyPath(): string;
    setPrivateKeyPath(value: string): DeriveKeyPairMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): DeriveKeyPairMessage.AsObject;
    static toObject(includeInstance: boolean, msg: DeriveKeyPairMessage): DeriveKeyPairMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: DeriveKeyPairMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): DeriveKeyPairMessage;
    static deserializeBinaryFromReader(message: DeriveKeyPairMessage, reader: jspb.BinaryReader): DeriveKeyPairMessage;
}

export namespace DeriveKeyPairMessage {
    export type AsObject = {
        keypairDetails?: NewKeyPairMessage.AsObject,
        publicKeyPath: string,
        privateKeyPath: string,
    }
}
// ...
export class KeyPairMessage extends jspb.Message { 
    getPublicKey(): string;
    setPublicKey(value: string): KeyPairMessage;

    getPrivateKey(): string;
    setPrivateKey(value: string): KeyPairMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): KeyPairMessage.AsObject;
    static toObject(includeInstance: boolean, msg: KeyPairMessage): KeyPairMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: KeyPairMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): KeyPairMessage;
    static deserializeBinaryFromReader(message: KeyPairMessage, reader: jspb.BinaryReader): KeyPairMessage;
}

export namespace KeyPairMessage {
    export type AsObject = {
        publicKey: string,
        privateKey: string,
    }
}